DM-433 update org edit team tab to correctly reflect SWU qualification status after removing a team member
