+++ src/front-end/typescript/lib/pages/organization/edit/index.tsx
@@ -4,20 +4,14 @@ import router from 'front-end/lib/app/router';
 import { SharedState } from 'front-end/lib/app/types';
 import * as TabbedPage from 'front-end/lib/components/sidebar/menu/tabbed-page';
 import { Immutable, immutable, mergePageAlerts, PageAlerts, PageComponent, PageInit, replaceRoute } from 'front-end/lib/framework';
-import * as api from 'front-end/lib/http/api';
 import * as Tab from 'front-end/lib/pages/organization/edit/tab';
 import Link, { routeDest } from 'front-end/lib/views/link';
 import React from 'react';
-import { doesOrganizationMeetSWUQualification, Organization } from 'shared/lib/resources/organization';
-import { isVendor, User, UserType } from 'shared/lib/resources/user';
+import { isVendor, UserType } from 'shared/lib/resources/user';
 import { adt, ADT, Id } from 'shared/lib/types';
 import { invalid, valid, Validation } from 'shared/lib/validation';
 
-interface ValidState<K extends Tab.TabId> extends Tab.ParentState<K> {
-  organization: Organization;
-  swuQualified: boolean;
-  viewerUser: User;
-}
+type ValidState<K extends Tab.TabId> = Tab.ParentState<K>;
 
 export type State_<K extends Tab.TabId> = Validation<Immutable<ValidState<K>>, null>;
 
@@ -37,34 +31,18 @@ function makeInit<K extends Tab.TabId>(): PageInit<RouteParams, SharedState, Sta
     userType: [UserType.Vendor, UserType.Admin],
 
     async success({ routePath, dispatch, routeParams, shared }) {
-      const organizationResult = await api.organizations.readOne(routeParams.orgId);
-      if (!api.isValid(organizationResult)) {
+      const params = await Tab.initParams(routeParams.orgId, shared.sessionUser);
+      if (!params) {
         dispatch(replaceRoute(adt('notFound' as const, { path: routePath })));
         return invalid(null);
       }
-      const organization = organizationResult.value;
-      const swuQualified = doesOrganizationMeetSWUQualification(organization);
-      const affiliationsResult = await api.affiliations.readManyForOrganization(organization.id);
-      if (!api.isValid(affiliationsResult)) {
-        dispatch(replaceRoute(adt('notFound' as const, { path: routePath })));
-        return invalid(null);
-      }
-      const viewerUser = shared.sessionUser;
       // Set up the visible tab state.
       const tabId = routeParams.tab || 'organization';
-      const tabState = immutable(await Tab.idToDefinition(tabId, organization).component.init({
-        organization,
-        swuQualified,
-        affiliations: affiliationsResult.value,
-        viewerUser
-      }));
+      const tabState = immutable(await Tab.idToDefinition(tabId, params.organization).component.init(params));
       // Everything checks out, return valid state.
       return valid(immutable({
-        organization,
-        swuQualified,
-        viewerUser,
         tab: [tabId, tabState],
-        sidebar: await Tab.makeSidebarState(organization, tabId)
+        sidebar: await Tab.makeSidebarState(params.organization, tabId)
       })) as State_<K>;
     },
     async fail({ routePath, dispatch, shared, routeParams }) {
@@ -81,7 +59,7 @@ function makeInit<K extends Tab.TabId>(): PageInit<RouteParams, SharedState, Sta
 }
 
 function makeComponent<K extends Tab.TabId>(): PageComponent<RouteParams, SharedState, State_<K>, Msg_<K>> {
-  const idToDefinition: TabbedPage.IdToDefinitionWithState<Tab.Tabs, K, ValidState<K>> = state => id => Tab.idToDefinition(id, state.organization);
+  const idToDefinition: TabbedPage.IdToDefinitionWithState<Tab.Tabs, K, ValidState<K>> = state => id => Tab.idToDefinition(id, state.tab[1].organization);
   return {
     init: makeInit(),
     update: updateValid(TabbedPage.makeParentUpdate({
@@ -93,11 +71,11 @@ function makeComponent<K extends Tab.TabId>(): PageComponent<RouteParams, Shared
     getAlerts: getAlertsValid(state => {
       return mergePageAlerts(
         {
-          info: !state.swuQualified && isVendor(state.viewerUser) && state.tab[0] !== 'qualification'
+          info: !state.tab[1].swuQualified && isVendor(state.tab[1].viewerUser) && state.tab[0] !== 'qualification'
           ? [{
               text: (
                 <div>
-                  This organization is not qualified to apply for <em>Sprint With Us</em> opportunities. You must <Link dest={routeDest(adt('orgEdit', { orgId: state.organization.id, tab: 'qualification' as const }))}>apply to become a Qualified Supplier</Link>.
+                  This organization is not qualified to apply for <em>Sprint With Us</em> opportunities. You must <Link dest={routeDest(adt('orgEdit', { orgId: state.tab[1].organization.id, tab: 'qualification' as const }))}>apply to become a Qualified Supplier</Link>.
                 </div>
               )
             }]
@@ -113,7 +91,7 @@ function makeComponent<K extends Tab.TabId>(): PageComponent<RouteParams, Shared
       getTitleSuffix: state => {
         return state.tab[0] === 'organization'
           ? 'Edit Organization'
-          : `${state.organization.legalName} — Edit Organization`;
+          : `${state.tab[1].organization.legalName} — Edit Organization`;
       }
     }), makePageMetadata('Edit Organization'))
   };
